Feat: hide switch network UI if not connected to Metamask wallet
